# api_server/docs/swagger.yaml
basePath: /api
definitions:
  errors.APIError:
    properties:
      code:
        $ref: '#/definitions/errors.APIErrorCode'
      error:
        type: string
    type: object
  errors.APIErrorCode:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - ErrCodeMissingParam
    - ErrCodeIncorrectParam
    - ErrCodeSQLNoRows
    - ErrCodeSQLInternalError
  types.APIResponse:
    properties:
      error:
        $ref: '#/definitions/errors.APIError'
      message:
        type: string
    type: object
  types.Client:
    properties:
      client_id:
        type: integer
      client_name:
        type: string
      is_admin:
        type: boolean
      last_edited:
        type: string
    type: object
...
  types.Payment:
    properties:
      client_id:
        type: integer
      last_edited:
        type: string
      payment_amount:
        type: number
      payment_date:
        type: string
      payment_id:
        type: integer
      room_id:
        type: integer
    type: object
  types.Room:
    properties:
      client_id:
        type: integer
      last_edited:
        type: string
      room_area:
        type: number
      room_id:
        type: integer
      room_people_count:
        type: integer
    type: object
info:
  contact: {}
  title: HACS database API
  version: "1.0"
paths:
...
  /client/id/{id}:
    delete:
      description: Delete client by telegram ID
      parameters:
      - description: Client telegram ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Delete client
      tags:
      - client
    get:
      description: Get client by telegram ID
      parameters:
      - description: Telegram ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/types.Client'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get client by telegram ID
      tags:
      - client
    patch:
      description: Patch client by client_id
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client name
        in: formData
        name: client_name
        type: string
      - description: is admin
        in: formData
        name: is_admin
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Patch client
      tags:
      - client
    post:
      description: Create new client
      parameters:
      - description: Client telegram ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client name
        in: formData
        name: client_name
        required: true
        type: string
      - description: Is admin
        in: formData
        name: is_admin
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: New client
          schema:
            $ref: '#/definitions/types.Client'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Create new client
      tags:
      - client
...
  /payment/room/id/{id}:
    get:
      description: Get all payments by room_id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Payment'
            type: array
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get all payments by room_id
      tags:
      - payment
  /room/all:
    get:
      description: Get all rooms from MySQL
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Room'
            type: array
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get all rooms
      tags:
      - room
  /room/client/id/{id}:
    get:
      description: Get rooms by client_id
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/types.Room'
            type: array
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get rooms by client_id
      tags:
      - room
  /room/id/{id}:
    delete:
      description: Delete room by room_id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Delete room by room_id
      tags:
      - room
    get:
      description: Get room by room_id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/types.Room'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: No rows
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get room by room_id
      tags:
      - room
    patch:
      description: Patch room by room_id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client ID
        in: formData
        name: client_id
        type: integer
      - description: Room area
        in: formData
        name: room_area
        type: number
      - description: People living in room
        in: formData
        name: room_people_count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "404":
          description: Record not founded in DB
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Patch room
      tags:
      - room
    post:
      description: Create new room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client ID
        in: formData
        name: client_id
        required: true
        type: integer
      - description: People living in room
        in: formData
        name: room_people_count
        required: true
        type: integer
      - description: Room area
        in: formData
        name: room_area
        required: true
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Incorrect parameter
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Create new room
      tags:
      - room
produces:
- application/json
swagger: "2.0"